scripts/requirements-vertex.txt
# Vertex AI Training Requirements
# Python 3.10+

# Core ML libraries
torch==2.1.2
transformers==4.37.2
datasets==2.16.1
accelerate==0.26.1
peft==0.8.2
bitsandbytes==0.42.0

# Model dependencies
sentencepiece==0.1.99
protobuf==4.25.2

# Google Cloud
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.42.1

# Utilities
tqdm==4.66.1
tensorboard==2.15.1
